Use Audacious DBus interface to toggle repeat/shuffle
The MPRIS interface auf Audacious is broken since, at least, 2.1. Toggling
repeat and shuffle by the Audacious DBus interface works in 1.5 as well as
in 2.2alpha.

Update issue 17
Status: Fixed
Audacious' DBus interfaces are almost fixed again:
http://jira.atheme.org/browse/AUD-55

The Audacious adapter commited with this revision works flawlessly with
Audacious 1.5 and 2.2alpha. Everything in between may cause issues which can
not be fixed within Remuco. So either downgrade to 1.5 or use the most recent
version of Audacious.