Don't use preprocessing in client/common
This just complicates the build process for the Android client while it
is mainly used by the JavaME client. The drawback now is that the
version has to be set manually and that the logging tool does not know
if running in emulation mode. I think this is not a big problem.