Initial Android client implementations
Update issue 34
There is an intial Android client which already communicates with a
player adapter if the client runs inside an emulator. This initial
client version is a kind of starting point for the Android client which
shows how things principally work in the context of Remuco.

This is my first Android code and it is not meant to be productive. It
is only intended to demonstrate the connection setup to and interaction
with a remote player (adapter).