Fix Exaile library search
Update issue 76
Status: Fixed
Labels: +Component-Exaile
The plugins's crash on an empty search query has been fixed.

Additionally an empty search now returns all tracks in the library.
Originally the intention was to return no tracks on an empty search but
as there is no other quick way to get a list of all tracks in the Exaile
adapter (except openening the smart playlist 'Entire Library' and
browsing that open playlist) returning the whole library on an empty
search seems reasonable.