Android: fix library browsing issue due to bogus client info
When connecting explicitly using the connect dialog, the client info has
not been intialized explicitly which caused an eception in the server
(because of a list page size of 0). This change fixes this issue.
However, the different ways to connect should be unified in order to not
handle client info intitialization twice (see also the added comments in
RemucoActivity.java).