MIDP: include client variant for Motorola K1 in release tarball
Update issue 84
Status: Fixed

The release tarball now contains an additional pre-built client variant
named 'motorola-k1-fix'. It does not call the flush method on output
streams - this method seems to be buggy on the named device. Note that
this fix theoretically might delay the sending of messages to the server.